tests/save-load: Test channel allocation policy
Three tests are added for the complex load configuration.

The first test checks that the absence of the node `<allocation_policy>'
in the configuration yields a channel with PER_CPU allocation.

The second test checks that setting the allocation policy to PER_CPU
yields a channel with PER_CPU allocation.

The third test checks that setting the allocation policy to PER_CHANNEL
yields a channel with PER_CHANNEL allocation.

Change-Id: I16f7bb18e15ef9705e4ef6164687dc5b7355b4d8
Depends-on: lttng-ust: I05107e08106b1dcd69dbde5bd76b9ff6ed9b15db
Signed-off-by: Olivier Dion <odion@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>