Add "cpu_id" context
The "cpu_id" context is implicitly enabled for the kernel domain. It is
also implicitly provided by UST domains, if the channel has the
allocation policy "per-cpu".  In that case, the context is still
registered internally by the session daemon, but not sent to UST to
avoid duplication of the context.

Change-Id: Iab67b315b282c81784e316c97a10de79cd4f25fb
Depends-on: lttng-ust: I05107e08106b1dcd69dbde5bd76b9ff6ed9b15db
Signed-off-by: Olivier Dion <odion@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>