extras/zsh-completion/_lttng: add `--buffer-ownership` completion
This patch adds Zsh completion for the LTTng 2.14+ `--buffer-ownership`
option of the `enable-channel` command.

This option is mutually exclusive with the `--buffers-*` options as they all
mean the same thing.

The choices are `system`, `user`, and `process` when there's no current
tracing domain option. With `--kernel`, you only get `system`. With
`--userspace`, you only get `user` and `process`.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Change-Id: Ia45a245369b88d1e7bf61400c131be5933bb22ae