doc/man: document the buffer ownership model and allocation policy
This patch adds documentation for the 2.14+ `--buffer-ownership` and
`--buffer-allocation` options of lttng-enable-channel(1).

The `--buffers-global`, `--buffers-uid`, and `--buffers-pid` options are
marked as deprecated since the `--buffer-ownership` option is
now equivalent.

The patch also updates the "Buffering scheme" section of
lttng-concepts(7) to match the new concepts. A new table indicates the
behaviour of LTTng depending on the values of the new
`--buffer-ownership` and `--buffer-allocation` options.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Change-Id: I25843d7a347d96d9d066f93714eb465f9ef7aea0