lttng-sessiond(8): remove `LTTNG_UST_ALLOW_BLOCKING` documentation
lttng-ust(3) already documents the purpose of this environment variable.

Documenting the usage of such a variable for the purpose of tracing the
session daemon itself isn't relevant to the end user and belongs to
internal documents, therefore I'm removing it (like I removed
`LTTNG_UST_APP_PATH` for the same reason).

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Change-Id: I8b9256129855b5bedc29ce56b0f76f9ce9f2f0d0