Tests: fix: time TAP tests in OOT builds
The location of the 'clock' utility is determined relative to the
tap-driver.sh script which resided in the source dir. However the
'clock' utility is a compiled binary and will thus reside in the build
dir which differs from the source dir in out-of-tree builds.

To work around this, copy tap-driver.sh to the build dir like we do for
other test scripts and execute it from there.

Change-Id: I1bbcce7ad0da9f4aff51a41d5368bb4306a4934e
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>