doc/api/liblttng-ctl/Makefile.am: depend on all source files
This patch adds dependencies to the `$(DX_DOCDIR)/$(PACKAGE).tag`
target, which is a tag file which the Doxygen process generates when
done.

The added dependencies are `Doxyfile` (the configuration), `style.css`
(custom style), everything under `dox` (general documentation), and all
the public liblttng-ctl headers.

The effect is that changing any of those files make the `doxygen-doc`
target regenerate the whole API documentation project.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: Idb5378126c9bc44b7e1b7c86333e4d0dce518edb
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>