sessiond: check liblttng-ust-ctl.so version at runtime
This commit introduces a runtime check to ensure that the version of
liblttng-ust-ctl.so loaded at runtime matches the version against which
LTTng-tools was built. This check is performed when the session daemon
is started, and if a mismatch is detected, an error message is logged
and the program exits with an error code.

This change is intended to prevent potential issues that may arise from
using incompatible versions of the library, which could lead to
undefined behavior or crashes.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Change-Id: Iea0d13aaab0f00e8d6124b66cc5e5ff959404d88