Tests: Fix hang in `tools/health/test_tp_fail`
With recent changes in utils.sh, `start_lttng_relayd` waits for the
PID file of daemon to be created. Failing to create the file results
in a hang when not running in background mode, and a failure when
running in background mode.

Use no-tap and background mode to handle both issues when starting the
relayd daemon during this test, as some of the test points may cause
the relayd to exit quickly.

Change-Id: I0b56df0d9131a828e1db855ae7fac1971cdd0399
Signed-off-by: Kienan Stewart <kstewart@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>