tests: Add test for hangs during high latency live sessions
==== Drawbacks ====

This test uses Python's implementations of setns and unshare, which
were introduce in 3.12. Testing with Python < 3.12 will skip the test.

This test assumes that the following netns do not already exist: ns1,
ns2.

This test assumes that the following ethernet devices do not already
exist: veth1, veth2.

The choice of a 30 second timeout is relatively arbitrary. On my
system with 4t, it passes. In a system with 192t, it could very well
not be sufficient as many more streams need to be flushed during a
live timer.

==== References ====

[1]: https://bugs.lttng.org/issues/1421

Change-Id: I074107e274e8bb02bad00688797590973359901c
Signed-off-by: Kienan Stewart <kstewart@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>