sessiond/consumerd: make the error protocol message-based
The error socket between the consumer and session daemon is used to
announce error codes -- values from the lttcomm_return_code enum -- from
the consumer daemon to the session daemon. As such, the error codes are
sent and received directly on the wire.

Since a follow-up change seeks to re-use the socket to transmit error
handling commands through the same socket, the protocol is changed to
become "message-based".

As such, every message sent from the consumer daemon(s) to the session
daemon through this socket now begin with the following header:

struct lttcomm_consumer_error_msg_header {
	/* enum lttng_consumer_error_msg_type */
	uint8_t msg_type;
	/* Size of payload following this member. */
	uint64_t size;
} LTTNG_PACKED;

That header is then followed by a message type-specific payload. As of
this change, the only message type that is supported is
"LTTNG_CONSUMER_ERROR_MSG_TYPE_ERROR_CODE, which has the following
layout:

struct lttcomm_consumer_error_msg_error_code {
	/* enum lttcomm_return_code */
	uint8_t error_code;
} LTTNG_PACKED;

The send and receive paths are adapted to expect this format and
pack/unpack it as needed.

No user-visible change is intended beyond the formatting of the log
messages.

Co-authored-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Signed-off-by: Olivier Dion <odion@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Depends-on: lttng-ust: I806f0f8b9143640efc1f7b3d864665736299a854
Change-Id: I067a83fc741c5319e7aa59c75f563866dd78c015