sessiond: allocate a unique application owner id on application creation
In order to make it possible for the user space tracer to uniquely
identify an application that was taking part in a ring-buffer
transaction, the new LTTng-UST APIs require a unique "owner" identity to
be supplied when a channel is created.

In preparation for this change, a unique "owner" identity is allocated
when an application registers. To support 32-bit applications, this
identity must be 32-bit wide, forcing us to plan for the recycling of
identities.

A new hash table, `ust_app_ht_by_owner_id` is added to those already
global in `ust-app.cpp`. Like `ust_app_ht_by_sock` and
`ust_app_ht_by_notify_sock`, it has no ownership of the application in
the sense that the lifetime of an entry in `ust_app_ht_by_owner_id_` is
bounded by the lifetime of its `ust_app` instance, itself ultimately
owned by `ust_app_ht`.

The allocation of owner identities is simply performed with a monotonic
counter, but the `ust_app_ht_by_owner_id` is checked to see if the id is
still in use.

Note that LTTng-UST reserves a number of owner identities:

- `LTTNG_UST_ABI_OWNER_ID_UNSET` is a reserved value to indicate the
"unset" state

- `LTTNG_UST_ABI_OWNER_ID_CONSUMER` is reserved to identify the consumer
daemon (e.g., when flushing)

Co-authored-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Signed-off-by: Olivier Dion <odion@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Depends-on: lttng-ust: I806f0f8b9143640efc1f7b3d864665736299a854
Change-Id: Id8a8886b049e4004fea1b5de219cbd8f30e170d0