Reclaim owner id on application teardown
From the sessiond-daemon's perspective, when a user application dies,
its owner-id must be reclaimed. To do so, it gets a list of the channel
keys used by that application and sends a message to the consumer-daemon
stating:

  For each channel matching these keys, check that this owner-id is not
  present in any sub-buffer of their corresponding streams.

But before doing so, it puts the owner-id in a garbage table with a
reference count equal to the number of channels previously listed --
making it an invalid identifier -- and then remove the owner-id from the
global owner-id table.

From the consumer-daemon perspective, a request of reclamation is
received. It iterates over the list of keys and dispatch to the
corresponding channels the owner-id to reclaim by adding it to a "owners
to be reclaimed" sets of each channel. The consumer-daemon replies to
this request with a single number stating how many channel were not
successfully notified. This tells the session-daemon if the reference
count in the garbage table needs to be decremented.

This commit does not introduce the fixing itself and so the channels and
the garbage table just accumulate the owner IDs until exhaustion.

Co-authored-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Signed-off-by: Olivier Dion <odion@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Change-Id: I0b28a8bede7f883760bfe1d55c6d1b03cd29132f
Depends-on: lttng-ust: I806f0f8b9143640efc1f7b3d864665736299a854