Add stall-watchdog timer to user space channels
The watchdog-timer is a timer associated with a user space channel. When
triggered, a scan and possibly a fixup of the sub-buffers of each stream
of the channel is made.

The algorithm works in two phases. Scanning the sub-buffers of each
stream to list the owners seen, first. Fixing up the sub-buffers using
the previous set of owners seen and the set of owners to be reclaimed
from the channel, second.

The scan works by iterating over the sub-buffers of a stream between the
consumer position and the producer position. The state of each
sub-buffer is polled using UST-ctl. It is compared with the last seen
state of that sub-buffer. If the state is the same, a warning is emitted
stating that there is a potential stalled buffer. In any case, the
owner-id in the state is added to a set of seen owners.

The fixup works by requesting UST-ctl to do a fixup on each stream,
using the previous set of owners seen and the intersection of owners to
be reclaimed by the channel.

If for any reasons a fixup failed for any of the streams, no owner IDs
are reclaimed. This is for conservative reasons and to simplify the API
of UST-ctl. If all fixup has succeed, then the channel can safely
reclaim all of the IDs to be reclaimed at the start of the algorithm.

The reclamation process is done by removing the owner IDs from the set
of to be reclaimed owner IDs of the channel and by notifying the
session-daemon with a single message containing a list of owner IDs that
were reclaimed.

From the session-daemon's perspective, when a reclamation response is
received, the corresponding owner IDs are looked-up in the reclamation
table and their reference counts are decremented. If a reference count
becomes zero, the corresponding owner ID is removed from the reclamation
table, making it a valid ID again.

Co-authored-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Signed-off-by: Olivier Dion <odion@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Change-Id: I4ee73a1469f5ad61ea9d52ddca2c51f801d74158
Depends-on: lttng-ust: I806f0f8b9143640efc1f7b3d864665736299a854