tests: Add UST stall tests
The stalls tests all work similarly.

A session is created with a channel and some events enabled. The session
is started.

Producers are spawned with a debugger and stopped at very precise
locations within the code of UST. This is done using test-points defined
in UST, which requires debug symbols.

When all producers are stopped at the desired locations, they are all
killed by the debugger and the session is destroyed.

The tests consist of ensuring that the destruction of the session ends
within some margin of time and that the resulting trace either matches
some criterion or is valid.

The simple tests involve a producer where the expected criterion are
explicitly written.

The complex tests involve multiple producers and the only criteria for
the final trace is that it is valid to be read with babeltrace.

Because the number of complex tests is huge, it is implemented with
multiprocessing in mind.

Signed-off-by: Olivier Dion <odion@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Change-Id: I91e01e8d6c79d2c60951e2e07f12be6f99ca2933
Depends-on: lttng-ust: I806f0f8b9143640efc1f7b3d864665736299a854