tests: tests.serial: polled writing and long-running test abort
Prior to this change, the serial test runner would not output anything
between the start of a test and the end. In the CI, activity on the
console is used as an indicator for hung tests. Therefore, a long
running test started by the serial runner could lead to an
false-positive abort.

This change reduces the polling timeout from 1s to a default of 60s,
writes the output to the test log at each interval, and logs an
informational message. This will prevent the false-positive aborts for
long running tests.

To avoid keeping a test run alive, the serial test runner keeps track
of the last timestamp that a non-zero amount of bytes from the
sub-processs has been written and will kill the subprocess if it has
been longer than the timeout period in minutes (Default: 30).

Change-Id: Icc968a3714f1a76fb1015ee55202106e270dca77
Signed-off-by: Kienan Stewart <kstewart@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>