Import of task scheduler and tests
Import the task scheduler from the 2023 NorthSec conference badge
project [1].

It is lightly modified to fit in the LTTng-tools project's structure.
The tests were also adapted to remove their dependency on the Unity test
framework.

Modifications were kept to a minimum to start from a "known-good" state
and to keep a history of changes in the tree.

It will be used in upcoming changes to manage periodic tasks.

As the original author, I am relicencing this code to LGPL 2.0.

[1] https://github.com/nsec/badge-conf-2023/tree/nsec2023

Change-Id: Ibbc7b58f38ee1051c24ecef85a1a7ccb915b27d7
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>