task scheduler: remove fixed maximum number of tasks limitation
The task scheduler was initially developed for a platform where dynamic
memory allocations were impractical and in a context where the number of
concurrent tasks could be easily bounded.

These restrictions no longer apply in the context of the LTTng-tools
project. Therefore, the task array is now dynamically resized as new
tasks are inserted.

Change-Id: I3a488313a08669f2789687f7d3c8b5cbc46ad9cb
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>