consumer: Cool-down file-descriptors when no data is available
Some streams might have data ready for consumption. However, it's
possible that the the next sub-buffer in the stream is currently
stalled.

In theory, this will be fixed by the watchdog timer later. However, it's
possible for the watchdog-timer to be waiting on the stream mutex to fix
it, which is currently held by the polling loop. Since pthread's
mutexes are not fair, if the polling loop has nothing else to do, it
will quickly release the stream lock, poll it's wake file-descriptor and
re-take the lock.  This creates a situation where the watchdog-timer is
starved by the polling loop, which is waiting for the watchdog-timer to
complete.

A possible fix for this is to add the problematic file-descriptors to a
cool-down set of file-descriptors. These file-descriptors are not
polled for one turn of the loop.

To keep fairness with the cool-downed file-descriptors, a timeout of 10 ms
is set when polling with a non-empty set of cool-downed file-descriptors.

Signed-off-by: Olivier Dion <odion@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Change-Id: I9d5f37dbefca15c6721758958835a5e964aab447
Depends-on: lttng-ust: I806f0f8b9143640efc1f7b3d864665736299a854