Fix: sessiond: stale ust_app_channel on initialization failure
Observed issue
==============

As part of an upcoming change which attempts to determine the channels
that are used by an application. To do so, an iteration over the
application's channels (ust_app_channel) is performed and their
corresponding registry channels are looked up in the session registry.

With those modifications in place, running the
test_per_application_leaks.py test fails as an unhandled
lttng::invalid_argument_error exception is thrown when calling
registry_session::channel() to look up the registry channel.

Cause
=====

The code rightfully assumes that the registry channel should be present
in the session registry. However, since the channel was not created
successfully on the consumer side, the registry channel corresponding
to the ust_app_channel instance isn't known to the session registry.

Solution
========

The publication of the ust_app_channel in the ust_app_session is delayed
until the channel creation on the consumer side is successful. This
ensures that the ust_app_channel is only published if the channel
creation was successful.

The ust_app_channel is deleted in case of failure to create the channel
on the consumer side. This is done in the error handling path of
ust_app_channel_create().

Known drawbacks
===============

None.

Change-Id: Ic44ef3581bec3176989bb9e79ec53fd52e8d5651
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>