common: introduce task-executor for scheduled task management
Add a task execution thread (and related tests) to "pump" the scheduler
and execute tasks as their deadline expires.

The task executor reacts to a timerfd and an eventfd:
  - The timerfd is used to wake the thread when the next task deadline
    is set to expire.
  - The eventfd is used when the thread must be woken up to either
    reevaluate the next upcoming deadline (e.g., when a new task is
    scheduled), or when the thread should exit.

The task scheduler is modified to use absolute times when specifying
when a task should be scheduled. This makes it easier to maintain tests
that don't depend on the system's clock while also retaining an
ergonomic API.

The lttng::synchro::waiter utility is moved to its own "synchro" library
as to not impose a dependency on liburcu in every user of libcommon-gpl.

Note that it was erroneously placed in the GPL-licensed library despite
being LGPL-licensed. This is now fixed.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Change-Id: I3798d673e34fc5265563a97a12398681f42fa809