tests: Validate that HOSTNAME is not empty
A number of tests use the bash built-in variable HOSTNAME. If this
variable is empty for any reason, the tests will fail and some may
hang indefinitely. While this may be indicative of a configuration
issue on the host running tests, the value can be checked early.

Fallbacks to hostname and uname are attempted, and the tests will
fail early if the value is still empty.

Change-Id: If68242cbd213ff3af8f987e0d03f6967f546668b
Signed-off-by: Kienan Stewart <kstewart@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>