consumerd: add a timer task executor thread
Add a scheduler and task executor to the lttng_consumer_local_data. To
do so, lttng_consumer_local_data is now allocated using the C++ runtime
facilites (new/delete).

The "sigprocmask" is initialized earlier to ensure it is set to block RT
signals before the task executor thread is launched
(on creation of the task_executor instance).

As of this change, no change in behaviour is intended beyond spawning a
new worker thread and tearing it down on shutdown.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Change-Id: Idf313f235d21c8db0be811cbd957012716eb6a52