Fix: lttng-sessiond: crash with kernel tracer during early shutdowns
Observed issue
==============

While running the following test with a root user and kernel tracer
available, the following crash can was observed:

    $ LTTNG_UST_CTL_PATH=/fake/fake/fake lttng-sessiond
    PERROR - 15:05:16.150645431 [Main]: bind: No such file or directory (in lttcomm_create_unix_sock() at unix.cpp:136)
    Error: Create unix sock failed: /fake/asdfasdffa/lttng-ust-sock-8
    lttng-sessiond: notification-thread-events.cpp:2228: int handle_notification_thread_command_remove_tracer_event_source(notification_thread_state*, int, lttng_error_code*): Assertion `source_element' failed.
    Aborted

Cause
=====

The initialization of the kernel tracer (`init_kernel_tracer`) is done
relatively early in the startup. This creates the tracer event notifier
group notification fd. Later, after the notifier thread and kernel
management threads arestarted this fd gets registered to the notifier.

In `cleanup_kernel_tracer`, it is assumed that if
`kernel_tracer_event_notifier_group_notification_fd` is not negative,
then the fd needs to be de-registered from the notification thread.

However, in cases where the startup of the sessiond stops after
`init_kernel_tracer` but before the registration of
`kernel_tracer_event_notifier_group_notification_fd`, the assert will be
triggered as there has never been a corresponding add.

Solution
========

Set a flag when the `kernel_tracer_event_notifier_group_notification_fd`
has been successfully registered to the notifier thread and use that in
conjuction with a non-negative value of
`kernel_tracer_event_notifier_group_notification_fd` to decide if the fd
should be de-registered.

Known drawbacks
===============

None.

Change-Id: Iec2ca4e2b7343a5c4e56f214f8854ff4eedb2feb
Signed-off-by: Kienan Stewart <kstewart@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>