ust-ctl: synchronize public header for preallocate_backing
preallocate_backing was introduced as part of
lttng_ust_ctl_consumer_channel_attr in LTTng-UST.

This change propagates the new version of the header to the internal
copy in LTTng-tools.

Note that `attr.preallocate_backing` is set to `true` in order
to preserve the current behaviour.

The internal copy of ust-ctl.h was also missing the owner_id field which
was recently introduced; it is added as part of the synchronization.

Depends-on: lttng-ust: I01f7a97479d13fc7033c50a8ad0b8d296b225a6d
Change-Id: I007dd5243cbda9bffa7c95ce0b448bdcea3c10c5
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>