Tests: ust/high_throughput: parallelize event counting
Most of the runtime of this test is spent counting events in the
resulting trace especially now that it uses twice the number of
CPUs to decide how many apps it should launch.

Informally, on my 168 core dev machine, this cuts down the time
spent decoding the trace from ~20 seconds to 1 second.

Change-Id: Ibf3128000807960d12d9d592b8fd12e36fea0618
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>