tests: notification: skip relevant tests when sessiond start fails
For unknown reasons the sessiond starts in this test were done without
tap output making it less obvious what was the source problem when I had
to debug test failures.

Add tap output to all start and stop of the sessiond and skip the
relevant tests when the sessiond fails to start.

Also some minor shellcheck fixes while we are here.

Change-Id: Ib869af99310401cca0f6a970d1ac42b33d9750b7
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>