sessiond: session: iterate on consumer channel keys by name
Extend the ltt_session `user_space_consumer_channel_keys` facility to
allow filtering on a channel name.

This facilitates the implementation of commands that target a
specific user-facing channel within a session. It makes it easy to
list the "keys" that identify consumerd-side channel instances
that result from this channel.

For instance, when the per-uid ownership model is used, a
recording session channel will result in the creation of distinct
per-uid channel instances in each user-space consumer daemon.

When the per-pid ownership model is used, a recording session
channel will result in the creation of a per-application instance
of the channel on the consumer daemon's end.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Depends-on: lttng-ust: I01f7a97479d13fc7033c50a8ad0b8d296b225a6d
Change-Id: I09c9c8d8133a1d2426d8a9324e5cc18cf9ca4394