sessiond: don't rely exclusively on file permissions for auth
The current permission check logic for commands assumes that the
sessiond client socket has proper permissions set to make it
inaccessible to users that shouldn't be able to run commands against the
session daemon.

As an additional precaution, explicitly check the user's credentials and
membership of the tracing group when a connection is established.

While not a fix, this makes the permission checking more resilient in
the face of misconfigurations.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Depends-on: lttng-ust: I01f7a97479d13fc7033c50a8ad0b8d296b225a6d
Change-Id: I0ff6f511b6b5885a6daf9f6635fa13de835f0c2a