fix: link error in libpause_(consumer|sessiond).so
Running the test suite in the package build process of both Debian 13
and Fedora 43 results in a test failure related to an undefined symbol
in the LD_PRELOAD libraries 'libpause_consumer.so' and
'libpause_sessiond.so', like so:

  /bin/sh: symbol lookup error: lttng-tools/tests/regression/tools/notification/.libs/libpause_sessiond.so: undefined symbol: _ZN3fmt3v117vformatB5cxx11ENS0_17basic_string_viewIcEENS0_17basic_format_argsINS0_7contextEEE

I haven't found why we don't see a similar behavior in CI an manual
testing but the easy fix is to simply link with the libfmt convenience
library.

This minimal change is specific to stable-2.14, I have proposed a more
intrusive rework of how we link with libfmt for the master branch.

Change-Id: I337c4469538d6603292195184005cdbd6d3f8e45
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>