sessiond: add memory reclamation command
Add the reclaim_channels_memory command to the session daemon. This
command will be used by the lttng-ctl interface handler (client.cpp)
to target a channel. It essentially build a list of stream groups
associated to a given user-facing channel and forward the command
to the user space consumer daemon(s).

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Depends-on: lttng-ust: Ida9fe9af426f6f615a280736671acbe520512a53
Change-Id: Ie46fb6a613e15abc870e38e3aeba99f27636c2e6