consumerd: add optional continuous reclamation of a channel
Introduce a new "continuously_reclaimed" parameter as part of the
consumer-side channel creation command. When enabled, the memory used by
sub-buffers is reclaimed as soon as they are consumed (provided they are
consumed continuously; not as part of a snapshot).

Note that following recent LTTng-UST changes, "getting" the next
sub-buffer makes it the "reader" sub-buffer. Previously, that was only
done for channels configured to use the "overwrite" event record loss
mode.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Depends-on: lttng-ust: Ida9fe9af426f6f615a280736671acbe520512a53
Change-Id: I9e594efeb136231cfc9ea688dc7f53bf16bfcbfe