liblttng-ctl: expose the channel preallocation policy setting
A channel's preallocation policy allows the user to control when a
stream's subbuffers are allocated. The current behaviour maps to the
"PREALLOCATE" policy which sees the consumer daemon zero-ing the
ring-buffer's backing files when allocating a stream.

The new policy, "ON_DEMAND", leaves the buffers unallocated. They are
allocated "on demand" when traced applications access the ring-buffer's
pages for the first time.

Note that the kernel space tracer doesn't support the "ON_DEMAND"
preallocation policy.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Depends-on: lttng-ust: Ida9fe9af426f6f615a280736671acbe520512a53
Change-Id: Ib77525dd3407f55b1d97fd8e95a21fffa69ec626