liblttng-ctl: expose the channel automatic reclamation policy setting
Expose the auto-reclamation policy through the public interface.
The auto-reclamation policy allows the user to specify the frequency at
which buffers should be reclaimed.

Note that the memory reclamation policy is applied at a sub-buffer
granularity.

The behaviour of the automatic memory reclamation policy differs based
on the event record loss mode of the channel.

Discard mode:
  - Sub-buffers that have been written-to more than "older_than_us"
  microseconds ago and _that have been consumed_ are reclaimed.

Overwrite mode:
  - Sub-buffers that have been written-to more than "older_than_us"
  microseconds ago are reclaimed.

  A consequence of this is that recorded events may be lost when
  automatic reclamation is used on a channel configured in this mode.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Depends-on: lttng-ust: Ida9fe9af426f6f615a280736671acbe520512a53
Change-Id: Iabbbcd3807c89af971dbc536d455731e4903a25d