liblttng-ctl: add channel stream-info query command
The channel stream-info query command allows liblttng-ctl to report the
properties of the various streams associated with a named recording
channel.

A set of data streams is a group of streams that results from the
configuration of a recording channel. For instance, a channel configured
to use the per-pid ownership model will result in a group of streams per
traced application.

For the moment, this API reports the data stream sets and provides a
snapshot of the memory use of the various streams.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Depends-on: lttng-ust: Ida9fe9af426f6f615a280736671acbe520512a53
Change-Id: Ie56c665c0dcffe32e01862147ed34d97e297ef7f