conditions/buffer-usage: Fix range checking on domain type value
Observed issue
==============

When creating a buffer usage trigger for the log4j2 domain, the
following error is emitted:

    $ lttng add-trigger --condition=buffer-usage -b 'low' -d 'log4j2' \
    -r 0.5 -s example_session -c example_channel --action notify
    Error: Failed to register trigger: Unknown error

In the lttng-session logs:

    DBG1 - 10:08:21.696772301 [Client management]: Accepting client
    command ... (in thread_manage_clients() at client.cpp:2522)
    Error: Invalid domain type value (6) found in condition buffer
    Error: Failed to register trigger: Protocol error occurred.

Cause
=====

The range checking in `init_condition_from_payload` is done by
validating that the domain_type value is within the enum's range using
`LTTNG_DOMAIN_NONE` as the lower bound and `LTTNG_DOMAIN_PYTHON` as
the upper bound.

When `LTTNG_DOMAIN_LOGJ2` was introduced, this code wasn't updated to
the new higher bound.

Solution
========

Add `LTTNG_DOMAIN_NR` to `enum lttng_domain_type` to serve as the end
of enum value, and check against that instead of
`LTTNG_DOMAIN_PYTHON`.

Known drawbacks
===============

None.

Change-Id: Ic0e3ed4de46ffa767247a13da97fa0899f0ecb20
Signed-off-by: Kienan Stewart <kstewart@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>