Fix: sessiond: assertion fails when looking up stream set
Iterating over the stream info sets of session's channels causes an
assertion to hit if no stream set have been allocated for a given
channel yet for a per-UID session.

This assertion was justified in the past since this iteration utility
was only used after the allocation of at least one stream set. However,
it is no longer the case as it is used when measuring the memory usage
of the various data streams. That can happen before a session has been
started, resulting in the assertion failing.

The problem can be reproduced by running
./tests/regression/tools/trigger/start-stop/test_start_stop after the
implementation of the data stream memory usage reporting command in the
session daemon.

Change-Id: Id1dcac105a8b20707af6a50644eebf3ed41075af
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>