consumerd: introduce a util to iterate on a channel's streams
Streams can be iterated-on in various ways in the consumer daemon.
To iterate on a given channel's stream, a user must either use the
channel's 'streams' linked list or the global stream_per_chan_ht.

The one to use depends on whether or not the channel has an active
output or not (i.e. it is a channel within a "snapshot") session.

The `stream_set` util hides this detail and provides the streams
from both "places" allowing the calling code to be simplified.

Change-Id: I331422f608e42498154f04f9c8ba2a819ef18482
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>