tests: tap-driver.sh: disable input buffering with mawk
Contrary to other AWK implementations, mawk buffers its input by default
which results in the TAP driver receiving the whole output of a test
only as it exits. This can be problematic for inactivity timeout checks
based on the output as well as inconvenient when interactively following
a test run.

Mawk supports a non-standard 'interactive' mode which disables input and
output buffering, check if the configured AWK binary supports it and
enable it.

Change-Id: Ida10dfec8a552012fa411c623acc438debdaca1d
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>