tests: Add regression check to validate quiet stream timestamps
While investigating a failure in
`tests/regression/tools/live/test_early_inactive_app.py` caused by a
proposed patch[1], I noted that the patch also subtly changed the
behaviour of the begin and end timestamps for packets on quiet
streams.

When a stream is created, the header of the first packet is
immediately initialized and the the begin timestamp set to the
creation time. In early versions of the proposed patch[2], the begin
timestamp was not being set correctly, which caused the begin and end
timestamps to match exactly.

This regression test validates that timestamps for a stream that is
created but has no events produced the begin and end timestamps have
at least a 1s separation in order, so that readers can infer there
were no events during that period.

[1]: https://review.lttng.org/c/lttng-ust/+/14961
[2]: https://review.lttng.org/c/lttng-ust/+/14961/2

Change-Id: Ib12a060aa9b9962fd065ea563d0dd96fd3a229f5
Signed-off-by: Kienan Stewart <kstewart@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>