Fix: sessiond: crash while running clear test
Issue observed
--------------

When LTTng-tools is built with clang++,
./tests/regression/tools/clear/test_ust.py fails on a crash of the
session daemon.

  #0  0x00007f9e43ea774c in ?? () from /usr/lib/libc.so.6
  #1  0x00007f9e43e4ddc0 in raise () from /usr/lib/libc.so.6
  #2  0x00007f9e43e3557a in abort () from /usr/lib/libc.so.6
  #3  0x00007f9e43e354e3 in ?? () from /usr/lib/libc.so.6
  #4  0x0000561e4fe94d41 in urcu_ref_put (ref=0x7f9e2c02ccd0, release=0x561e4fe94d60 <trace_archive_location_destroy_ref(urcu_ref*)>) at /usr/include/urcu/ref.h:64
  #5  0x0000561e4fe94c05 in lttng_trace_archive_location_put (location=0x7f9e2c02ccd0) at location.cpp:62
  #6  0x0000561e4fd290e5 in handle_notification_thread_command_session_rotation (state=0x7f9e401fb6e0, cmd_type=NOTIFICATION_COMMAND_TYPE_SESSION_ROTATION_COMPLETED, session_id=1, trace_archive_chunk_id=0, location=0x7f9dfc000ba0, _cmd_result=0x7f9e3e9f7d10)
    at notification-thread-events.cpp:2032
  #7  0x0000561e4fd25c90 in handle_notification_thread_command (handle=0x561e673aad00, state=0x7f9e401fb6e0) at notification-thread-events.cpp:3223
  #8  0x0000561e4fd207bd in thread_notification (data=0x561e673aad00) at notification-thread.cpp:674
  #9  0x0000561e4fd4c1a8 in launch_thread (data=0x561e673afdc0) at thread.cpp:67

The trace_archive_chunk location has a suspicious refcount:
  (gdb) p location->ref
  $1 = {refcount = -9223372036854775808}

Cause
-----

The notification_thread_command structure is used to describe a command
being delegated to the notification thread. It contains a union with all
possible parameters for the various commands (named `parameters`).

That union is initialized using an aggregate initialization using an
empty initialized list (`= {}`).

Initializing a union in this manner is problematic. Indeed, when you use
`= {}` on a union, the C++ standard specifies that only the first member
is value-initialized (zero-initialized). If other union members are
larger than the first member, the remaining bytes of the union are left
uninitialized.

Solution
--------

The two unions of the structure are zero-initialized using memset()
within the constructor. The init_notification_thread_command() is also
no longer needed since the structure now defines a proper constructor.

Change-Id: Id9444622500792e6b74ca49eaab2e3624102cdd6
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>