relayd: Update stream inactivity when receiving live beacons
Observed issue
==============

While working on a proposed patch[1], it was noted that the change
caused the `tests/regression/tools/live/test_early_inactive_app.py`
regression test to hang and fail.

Cause
=====

In the proposed patch streams would not automatically have their
headers initialized. This means that there is no data to consume, and
the sessiond will correspondingly send a live beacon to indicate that
there is no activity.

The live beacon handling in lttng-relayd assumes that a stream must
have received at least one index in order to be marked as
inactive.

As the streams would never be marked as inactive, `check_index_status`
would be run and given that no index had been received would issue
`LTTNG_VIEWER_INDEX_RETRY`, causing live viewers to continue
requesting indexes for streams that have no activity.

Solution
========

Do not enforce the assumption that a stream must have at least one
index when handling live beacons in the relayd. When receiving a live
beacon, the stream will be marked as inactive unless there are indexes
currently in flight.

Known drawbacks
===============

None.

References
==========

[1]: https://review.lttng.org/c/lttng-ust/+/14961

Change-Id: I3eb5758d1ee6aef1a67a5de6bd21faae0bf33a03
Signed-off-by: Kienan Stewart <kstewart@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>